Refactor os specific commands to an 'os' module
