+++ src/commands/install/windows.rs
@@ -4,6 +4,7 @@ use semver::Version;
 
 use crate::{
     commands::{self, install::pip::install_extra_pip_packages},
+    os::windows::build_filename_exe,
     utils, Result,
 };
 
@@ -30,7 +31,7 @@ pub fn unattended_windows_install(
     ];
 
     let cwd = utils::pycors_download()?;
-    let exe = format!("./{}", utils::build_filename_exe(version)?);
+    let exe = format!("./{}", build_filename_exe(version)?);
 
     utils::run_cmd_template(
         &version,
